feat(cloudfront): Lambda@Edge construct
DRAFT PR - Looking for early-stages feedback

This PR creates a construct (`EdgeFunction`) for Lambda@Edge functions.
CloudFront requires that a function be in us-east-1 to be used with Lambda@Edge,
even if the logical distribution is created via another region. The initial goal
of this construct is to make it easier to request and work with a function in
us-east-1 when the primary stack is in another region.

In the future, this can be extended to validate other Lambda@Edge restrictions.
See https://docs.aws.amazon.com/AmazonCloudFront/latest/DeveloperGuide/lambda-requirements-limits.html
and #9833 for more information on those limitations.

Some open questions:
1. Is there a more clever way to "refresh" the SSM parameter reader when the
   underlying function changes?
2. How to make this and `edgeArn` play nicely together?

Thanks to @asterikx for the inspiration and consolidated writeup in #1575.

fixes #9862